apps/bot/pyproject.toml
[project]
name = "galamsey-tracker-bot"
version = "0.1.0"
description = "Telegram bot for citizen galamsey reports"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.6.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
